+++ docs/changelog.rst
@@ -118,6 +118,23 @@ BetterUptime
 ============
 
 *********
+``1.6.0``
+*********
+
+2021-05-28
+
+- Fix commands
+- Fix config migration
+
+*********
+``1.5.2``
+*********
+
+2021-05-25
+
+- Remove custom uptime command... There's some broken shit that I can't fix, rewrite was already planned and this will be fixed then (#23 on GitHub)
+
+*********
 ``1.5.1``
 *********